+++ packages/tron-wallet-snap/jest.config.js
@@ -19,10 +19,10 @@ module.exports = {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 72.42,
-      functions: 79.64,
-      lines: 85.79,
-      statements: 85.81,
+      branches: 72.63,
+      functions: 79.95,
+      lines: 85.84,
+      statements: 85.86,
     },
   },
 };